Updated for multi processing with data compression
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@9107 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ daq/storage/modules/include/StorageDeserializer.h
@@ -65,7 +65,7 @@ namespace Belle2 {
     int m_nodeid;
     std::string m_nodename;
     int m_shmflag;
-    unsigned int m_count;
+    unsigned long long m_count;
     unsigned int m_expno;
     unsigned int m_runno;
     unsigned int m_evtno;
